+++ src/lib/toast/Toast.svelte
@@ -1,13 +1,12 @@
 <script lang="ts">
-  import { createEventDispatcher, type ComponentProps } from 'svelte';
-  import Frame from '../utils/Frame.svelte';
+  import TransitionFrame from '$lib/utils/TransitionFrame.svelte';
+  import type { ComponentProps } from 'svelte';
   import { twMerge } from 'tailwind-merge';
   import CloseButton from '../utils/CloseButton.svelte';
-  import { fade } from 'svelte/transition';
-  import type { Dismissable } from '$lib/types';
+  import Frame from '../utils/Frame.svelte';
 
   // propagate props type from underying Frame
-  interface $$Props extends ComponentProps<Frame>, Dismissable {
+  interface $$Props extends ComponentProps<TransitionFrame> {
     color?: 'primary' | 'gray' | 'red' | 'yellow' | 'green' | 'blue' | 'indigo' | 'purple' | 'orange' | 'none';
     position?: 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right' | 'none';
     open?: boolean;
@@ -20,14 +19,10 @@
   export let dismissable: boolean = true;
   export let color: 'primary' | 'gray' | 'red' | 'yellow' | 'green' | 'blue' | 'indigo' | 'purple' | 'orange' | 'none' = 'primary';
   export let position: 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right' | 'none' = 'none';
-  export let open: boolean = true;
   export let divClass: string = 'w-full max-w-xs p-4 text-gray-500 bg-white shadow dark:text-gray-400 dark:bg-gray-800 gap-3';
   export let defaultIconClass: string = 'w-8 h-8';
   export let contentClass: string = 'w-full text-sm font-normal';
   export let align: boolean = true;
-
-  const dispatch = createEventDispatcher();
-  $: dispatch(open ? 'open' : 'close');
 
   const positions = {
     'top-left': 'absolute top-5 left-5',
@@ -36,11 +31,6 @@
     'bottom-right': 'absolute bottom-5 right-5',
     none: ''
   };
-
-  function close(e: MouseEvent) {
-    e.stopPropagation();
-    open = false;
-  }
 
   let finalDivClass: string;
   $: finalDivClass = twMerge('flex', align ? 'items-center' : 'items-start', divClass, positions[position], $$props.class);
@@ -62,27 +52,27 @@
   $: iconClass = twMerge('inline-flex items-center justify-center shrink-0', colors[color], defaultIconClass);
 
   const clsBtnExtraClass = '-mx-1.5 -my-1.5 text-gray-400 hover:text-gray-900 focus:!ring-gray-300 hover:bg-gray-100 dark:text-gray-500 dark:hover:text-white dark:hover:bg-gray-700';
+
+  $: console.log($$restProps);
 </script>
 
-{#if open}
-  <Frame rounded transition={fade} color="none" {...$$restProps} class={finalDivClass} role="alert">
-    {#if $$slots.icon}
-      <Frame rounded color="none" class={iconClass}>
-        <slot name="icon" />
-      </Frame>
-    {/if}
+<TransitionFrame rounded {dismissable} color="none" role="alert" {...$$restProps} class={finalDivClass} let:close on:close>
+  {#if $$slots.icon}
+    <Frame rounded color="none" class={iconClass}>
+      <slot name="icon" />
+    </Frame>
+  {/if}
 
-    <div class={contentClass}>
-      <slot />
-    </div>
+  <div class={contentClass}>
+    <slot />
+  </div>
 
-    {#if dismissable}
-      <slot name="close-button" {close}>
-        <CloseButton class={clsBtnExtraClass} on:click={close} />
-      </slot>
-    {/if}
-  </Frame>
-{/if}
+  {#if dismissable}
+    <slot name="close-button" {close}>
+      <CloseButton class={clsBtnExtraClass} on:click={close} />
+    </slot>
+  {/if}
+</TransitionFrame>
 
 <!--
 @component
