+++ src/routes/component-data/Badge.json
@@ -1,1 +1,1 @@
-{"name":"Badge","slots":["close-button"],"events":[],"props":[["color","'primary' | 'blue' | 'dark' | 'red' | 'green' | 'yellow' | 'indigo' | 'purple' | 'pink' | 'none'","'primary'"],["large","boolean","false"],["dismissable","boolean","false"]]}
+{"name":"Badge","slots":["close-button"],"events":["on:close"],"props":[["color","'primary' | 'blue' | 'dark' | 'red' | 'green' | 'yellow' | 'indigo' | 'purple' | 'pink' | 'none'","'primary'"],["large","boolean","false"],["dismissable","boolean","false"]]}
