+++ src/routes/component-data/TransitionFrame.json
@@ -1,0 +1,1 @@
+{"name":"TransitionFrame","slots":[],"events":[],"props":[["transition","TransitionFunc","fade"],["params","object","{}"],["open","boolean","true"],["dismissable","boolean","false"]]}
